+++ src/adapters/http.js
@@ -30,7 +30,31 @@ function DSHttpAdapterProvider() {
      */
     queryTransform: function (resourceName, params) {
       return params;
-    }
+    },
+
+    /**
+     * @doc property
+     * @id DSHttpAdapterProvider.properties:defaults.$httpConfig
+     * @name defaults.$httpConfig
+     * @description
+     * Default `$http` configuration options used whenever `DSHttpAdapter` uses `$http`.
+     *
+     * ## Example:
+     * ```js
+     * angular.module('myApp', function (DSHttpAdapterProvider) {
+     *   angular.extend(DSHttpAdapterProvider.defaults.httpConfig, {
+     *     interceptor: [...],
+     *     headers: {
+     *       common: {
+     *         Authorization: 'Basic YmVlcDpib29w'
+     *       }
+     *     },
+     *     timeout: 20000
+     *   });
+     * });
+     * ```
+     */
+    $httpConfig: {}
   };
 
   this.$get = ['$http', '$log', 'DSUtils', function ($http, $log, DSUtils) {
@@ -308,6 +332,7 @@ function DSHttpAdapterProvider() {
     function HTTP(config) {
       var start = new Date().getTime();
 
+      config = DSUtils.deepMixIn(config, defaults.$httpConfig);
       return $http(config).then(function (data) {
         $log.debug(data.config.method + ' request:' + data.config.url + ' Time taken: ' + (new Date().getTime() - start) + 'ms', arguments);
         return data;